훅 분리
작업중.. 차후 더 개선할 여지가 있음
