+++ src/hook/useOrderCompleteMessage.ts
@@ -8,6 +8,7 @@ import { useEffect } from "react";
 import { toast } from "react-toastify";
 import { orderAPI } from "@/utils/orderApi";
 import { useQuery } from "@tanstack/react-query";
+import useProductSummary from "./useProductSummary";
 
 
 function useOrderCompleteMessage() {
@@ -15,32 +16,9 @@ function useOrderCompleteMessage() {
   const { receivers } = useReceiver();
 
   const navigate = useNavigate();
-  const location = useLocation();
 
-  const query = new URLSearchParams(location.search);
-  const idParam = query.get('id');
-  const id = idParam !== null ? Number(idParam) : null;
-
-  const productUrl = `${baseUrl}/api/products/${id}/summary`;
-  const { data, error } = useQuery<ProductItemSummary>({
-    queryKey : ['productData'],
-    queryFn : () => getFromUrl(productUrl)
-  })
-  //const { item, error } = useFetchFromUrlT<ProductItemSummary>(productUrl, getFromUrl, defaultProductItemSummary);
-
-
-  useEffect(() => {
-    if (error) {
-      toast.error((error as Error).message);
-      navigate('/');
-    }
-  }, [error, navigate]);
-
-  const price = data?.price;
-  const imageUrl = data?.imageURL;
-  const name = data?.name;
-  const brandName = data?.brandName;
-
+  const total = receivers.reduce((acc, receiver) => acc + receiver.quantity, 0);
+  const {id, name } = useProductSummary();
 
   const handleOrder = async () => {
     ordererName.validate();
@@ -68,9 +46,9 @@ function useOrderCompleteMessage() {
     }
   };
 
-  const total = receivers.reduce((acc, receiver) => acc + receiver.quantity, 0);
+  
     return {
-        price, imageUrl, name, brandName, total, handleOrder
+      total, handleOrder
     }
 };
